expense-tracker-backend/requirements.txt
asgiref==3.8.1
boto3==1.35.95
botocore==1.35.95
Django==5.1.4
django-cors-headers==4.6.0
dynamodb-orm==0.2.0
jmespath==1.0.1
joserfc==0.9.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0

# Testing and linting dependencies
flake8==7.0.0
pytest==8.3.4
pytest-django==4.8.0
pytest-cov==4.1.0
coverage==7.4.0
moto==5.0.27
